Changed input specs for rates.
Reaction rates are now enclosed in a | <rate statement> | block,
where rate statement can either be a number or an expression
evaluating to a double at run time. Globally, any expression in
the input deck now has to be enclosed in curly braces and, if
allowed in the specific section, can contain multiple statements
separated by a colon.