Remove post-parsing variable consistency check
since we now check during parsing already. Since
we try to simplify all variables at parse time for
run-time efficiency, we can not allow molecule counts
to appear in variable definitions any more. We still
allow the TIME variable, with the caveat that it
will be evaluated to its value at time zero making
it kind of pointless.